+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Handle timezone offset changes on remote FAT filesystems
 - Drop reading the old state storage format
 - Do not mention my forked prim-ftpd, PR-s got merged in the original repo
+- Remove --valid-chars option
 
 ## 0.6.0